Отображение данных, полученных с сервера. Часть1
