+++ js/api.js
@@ -1,14 +1,17 @@
+import { createPosts } from "./data.js";
+
 export const getData = (onSuccess) => {
-  fetch('https://26.javascript.pages.academy/kekstagram')
+  fetch('https://26.javascript.pages.academy/kekstagram/data')
     .then((response) => response.json())
-    .then((wizards) => {
-      onSuccess(wizards);
+    .then((messages) => {
+      onSuccess(messages);
+      createPosts(messages);
     });
 };
 
 export const sendData = (onSuccess, onFail, body) => {
   fetch(
-    'https://26.javascript.pages.academy/kekstagram',
+    'https://26.javascript.pages.academy/kekstagram/data',
     {
       method: 'POST',
       body,
@@ -25,3 +28,11 @@ export const sendData = (onSuccess, onFail, body) => {
       onFail('Не удалось отправить форму. Попробуйте ещё раз');
     });
 };
+
+export function onSuccessMessage() {
+  console.log('Успешно');
+}
+
+export function onFailMessage() {
+  console.log('Неудачно');
+}
